pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    # Add other runtime dependencies here
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["app"]

[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "isort",
    # Add other development dependencies here
]
test = [
    "pytest",
    # Add your test dependencies here
]
prod = [
    "pandas"
    # Add your production dependencies here
]

[project.scripts]
app = "app.main:main"